dialectic/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dialectic"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "asyncpg==0.29.0",
    "httpx==0.26.0",
    "pydantic>=2.5.3",
    "pydantic[email]>=2.5.3",
    "python-multipart>=0.0.12",
    "websockets==12.0",
    "pyjwt>=2.9.0",
    "pwdlib[argon2]>=0.2.0",
    "exponent-server-sdk>=2.0.0",
    "anthropic>=0.25.0,<1.0.0",
    "tiktoken>=0.7.0",
    "openai>=1.12.0,<2.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]

[tool.setuptools.package-dir]
"" = "."

[tool.pytest.ini_options]
testpaths = ["tests"]
